fix: rename DAL to fetchAPI
