+++ src/pages/AllTracs/AllTracs.jsx
@@ -2,14 +2,14 @@ import * as S from './styles';
 import { SearchFilter, TracksList } from 'components';
 import { useEffect, useState } from 'react';
 import { setTracks } from 'store/tracksSlice';
-import { fetchTracks } from 'helpers/DAL';
+import { fetchTracks } from 'helpers/fetchAPI';
 import { useDispatch } from 'react-redux';
 
-export const AllTracs = ({loadingClass, setLoadingClass}) => {
+export const AllTracs = ({ loadingClass, setLoadingClass }) => {
   const dispatch = useDispatch();
   // Загружаю все треки
   useEffect(() => {
-    console.log('Запущент юзЭффект и класс ставится лоадинг')
+    console.log('Запущент юзЭффект и класс ставится лоадинг');
     setLoadingClass('loading');
     fetchTracks('all')
       .then((data) => {
@@ -26,7 +26,7 @@ export const AllTracs = ({loadingClass, setLoadingClass}) => {
     <S.Main>
       <S.Heading>Треки</S.Heading>
       <SearchFilter />
-      <TracksList playlist='list' loadingClass={loadingClass} />
+      <TracksList playlist="list" loadingClass={loadingClass} />
     </S.Main>
   );
 };
